Add Array.ID() and OrderedMap.ID()
Currently Array.StorageID() and OrderedMap.StorageID() are used as
identifier in client because storage IDs are guaranteed to unique.
However, storage ID should be only used to retrieve slabs (registers)
from storage.

In the future, when Atree register inlining is implemented,
some resources may not be stored in separate slabs, so they
will not have storage IDs anymore.

This commit adds ID() to uniquly identify Array and OrderedMap.  For now,
this is implemented as identical to StorageID in raw bytes ([16]bytes).
In the future, this can be changed to be decoupled from storage ID
completely.  Clients should use ID instead of StorageID to identify
atree composite values.