Refactor encoding version and flag to add 4 more flags
Currently, version number uses 1 byte and flags uses 1 byte.

We don't need 255 versions but we need more flags in order
to support atree inlining.

For now, reserve the first 4 bits for version, but also reserve
the 5th bit separately so in the future, we can decide if
we want to use 5th bit to add 16 more versions or a new flag.

So the first byte has (beginning with top bit):
1-4: version (0 to 15)
5:   reserved to provide 16 more versions or another flag
6:   reserved (use this next if needed)
7:   slab has next slab ID (only relevant for data slab)
8:   slab has inlined slabs (for atree inlining)

This change is backward compatible.