Uninline slab when it is overwritten or removed
Currently, Set() and Remove() return overwritten or removed storable.
If storable is inlined slab, it is not stored in storage (because it
is removed from its parent slab which is in storage), so any future
changes to it would be lost.

On the other hand, if overwritten or removed storable is SlabIDStorable,
any future changes to it can still be persisted because it is in
its own slab in storage.

This inconsistency (not merged or deployed yet) can cause potential
data loss or unexpected behavior if deployed.

This commit uninlines inlined slabs that is overwritten or removed,
stores them in storage, and returns their SlabID as storable to caller.