No-op on parentUpdater() if child isn't in parent
This commit checks if child is in parent container at adjusted index or
under the same key in parentUpdater() before modifying parent container.

If child is no longer part of parent, parentUpdater() returns with no-op,
and this callback is set to nil.

This is to handle outdated child reference modifying parent after it
is removed.