Add feature to fix references to non-existent register
The intended use case is to enable migration programs in onflow/flow-go to fix
broken references.  As of April 2024, only 10 registers in testnet (not mainnet)
were found to have broken references and they seem to have resulted from a bug
that was fixed 2 years ago by https://github.com/onflow/cadence/pull/1565.

A broken reference is a StorageID referencing a non-existent slab.

This commit adds FixLoadedBrokenReferences(), which traverses loaded slabs and
fixes broken references in maps.  To fix a map containing broken references,
new function replaces broken map with empty map having the same StorageID and
also removes all slabs in the old map.

Limitations:
- only fix broken references in map (this is intentional)
- only traverse loaded slabs in deltas and cache

IMPORTANT: This should not be used to silently fix unknown problems. It
should only be used by migration programs to fix known problems which
were determined to be appropriate to fix.