Add feature to fix refs to non-existent registers
In testnet, broken references seem to have resulted from a bug that was
fixed 2 years ago by https://github.com/onflow/cadence/pull/1565.

A broken reference is a `StorageID` referencing a non-existent register.
So far, only 10 registers in testnet (none on mainnet) were found to
contain broken references.

This commit adds a feature to enable migration programs in onflow/flow-go
to fix broken references in maps.

`FixLoadedBrokenReferences()` traverses loaded slabs and replaces
broken map (if any) with empty map having the same StorageID and
also removes all slabs in the old map.

Limitations:
- only fix broken references in map (this is intentional)
- only traverse loaded slabs in deltas and cache

IMPORTANT: This should not be used to silently fix unknown problems.
It should only be used by migration programs to fix known problems
which were determined to be appropriate to fix in this manner.