Update comment for NondeterministicFastCommit
Make it clearer that encoded slabs are still deterministic, so
array and map iterations will remain deterministic.

Only the sequence of changed slabs getting committed is
nondeterministic.

This is useful for migration programs that don't require
commit sequence of slabs to be deterministic while
still preserving deterministic encoding of slabs
(e.g. iteration of arrays and maps remain deterministic).