Support WrapperValue in Array.Set and Array.Remove
This commit adds support for the new WrapperValue interface in
Array.Set() and Array.Remove().

Overwritten and removed inlined slabs need to be uninlined and
stored in storage before returning to the caller.

Because of new interfaces of WrapperValue and WrapperStorable,
this commit uninlines wrapped inlined slabs that are overwritten
or removed.