Support WrapperValue in Map.Set and Map.Remove
This commit adds support for the new WrapperValue interface in
Map.Set() and Map.Remove().

Overwritten and removed inlined slabs need to be uninlined and
stored in storage before returning to the caller.

Because of new interfaces of WrapperValue and WrapperStorable,
this commit uninlines wrapped inlined slabs that are overwritten
or removed.