Move extracted subtests to EOF for readablity
Some subtests were extracted in order to allow more tests
to run in parallel.

However, it can be tedious to review on GitHub because the
diff shows additions and substractions intermingled.

This commit moves the extracted subtests to the end of file,
to make the diff easier to read by having these additions separated
from the subtractions.

This commit also restores some previously extracted functions for
better Git readability.  Maybe we can address this in a separate PR.